spine_flownet/bone_segmentation_utils/requirements.txt
albumentations==0.5.2
configargparse==1.2.3
matplotlib
monai==0.4.0
nibabel==3.2.1
numpy
opencv-python
pandas
pathlib
polyaxon-client==0.6.1
pytorch_lightning==1.2.5
scikit-learn
SimpleITK
torchgeometry==0.1.2
tqdm==4.51.0
scikit-image
python-math
wandb
geomstats